Run pdb with `python -m` when in a venv
Virtual environments do not offer the `pdb` executable, but we can still
run `pdb` by invoking `python -m pdb`.

In Emacs `M-x pdb` offers to run the usual `pdb` executable, so let's
change that when we invoke `venv-workon`. When we run `venv-deactivate`
we restore the original behavior.

This could have been implemented using `advice-add`, but unfortunately
that is not compatible with versions of Emacs <= 24.3 (where `defadvice`
should be used).

This commit addresses issue #26.