Set gud-pdb-command-name to an arbitrary value
We want to ensure that whenever we run `venv-workon` the
`gud-pdb-command-name` variable is set to `python -m pdb`, and when we
run `venv-deactivate` it's set back to whatever it was. We are not
really interested in the original value, so let's use an arbitrary one
so we won't have a void one.

This should also fix the Ci build.