Add support to auto-activate virtualenv found in projetile-project-root
Now it is possible to automatically activate a virtualenv located in the
root of a projectile project (projectile-project-root) however this
isn't done magical, the user has to call `(venv-projectile-auto-workon)'
in some point after entering the project, i.e. via projectile-hooks,
python-mode hooks or the action `projectile-switch-project-action' which
is the recommended way and thus documented in the README.md.

By default it will detect only virtualenvs whose name is ".venv" but
the user can expand the options by settings all the desired names in
the `venv-dirlookup-names' list, this is all documented in the README.